requirements.txt
watchdog
python-docx
pdfplumber
sentence-transformers
scikit-learn
joblib